chore: debug curve pool integration

+++ contracts/interfaces/ICryptoSwap.sol
@@ -31,8 +31,8 @@ interface ICryptoSwap {
 
     // Swap token i to j with amount dx and min amount min_dy
     function exchange(
-        uint256 i,
-        uint256 j,
+        int128 i,
+        int128 j,
         uint256 dx,
         uint256 min_dy
     ) external returns (uint256);
